Edit media messages (#557)
* Update

* Update

* Update

* Update (unstable)

* Update

* Update

* Update MediaViewController.java

* Update

* Update

* Update

* Update

* Fixes

* Remove Editor interface

* Replace icon + add copyright for created files

* Fullscreen fix

* Update

* Update

* Fix for strange bug?

* Revert "Fix for strange bug?"

This reverts commit 13cfb249f4d506adcab23ffe670bfd4b07238b8a.

* Fix

* Add "and X more winners"

* Create dot_baseline_file_replace_24.xml

* Update

* Update

* Update

* Update

* Update

* Update MessagesController.java

* Update MessagesController.java

* Update app/src/main/java/org/thunderdog/challegram/data/ContentPreview.java

---------

+++ app/src/main/java/org/thunderdog/challegram/widget/FileProgressComponent.java
@@ -390,6 +390,11 @@ public class FileProgressComponent implements TdlibFilesManager.FileListener, Fa
     this.mimeType = mimeType;
   }
 
+  @Nullable
+  public String getMimeType () {
+    return mimeType;
+  }
+
   public void setFile (@Nullable TdApi.File file) {
     setFile(file, null);
   }
@@ -783,6 +788,9 @@ public class FileProgressComponent implements TdlibFilesManager.FileListener, Fa
       }
       case TdlibFilesManager.STATE_IN_PROGRESS: {
         if (file != null) {
+          if (tdlib.cancelEditMessageMedia(chatId, messageId)) {
+            return true;
+          }
           if (file.remote.isUploadingActive || isSendingMessage) {
             tdlib.deleteMessagesIfOk(chatId, new long[] {messageId}, true);
           } else {
